requirements.txt
fastapi
uvicorn
langchain
langchain-community
faiss-cpu
requests
python-dotenv


